elastic-stack-plantas-industriales/elastic/elasticsearch/requirements.txt
certifi==2023.11.17
elastic-transport==8.12.0
elasticsearch==8.12.0
python-dotenv==1.0.1
urllib3==2.1.0
